keep size/threshold to uint16
There's no reason to hold uint256 internally. We can still pass uint256 to external keep dependency,
while using uint16 internally to save gas

+++ solidity/contracts/deposit/Deposit.sol
@@ -109,8 +109,8 @@ contract Deposit is DepositFactoryAuthority {
         address _TBTCDepositToken,
         address _FeeRebateToken,
         address _VendingMachine,
-        uint256 _m,
-        uint256 _n,
+        uint16 _m,
+        uint16 _n,
         uint64 _lotSizeSatoshis
     ) public onlyFactory payable returns (bool) {
         self.TBTCSystem = _TBTCSystem;
