+++ solidity/contracts/test/deposit/TestDeposit.sol
@@ -14,8 +14,8 @@ contract TestDeposit is Deposit {
         address _TBTCDepositToken,
         address _FeeRebateToken,
         address _VendingMachine,
-        uint256 _m,
-        uint256 _n,
+        uint16 _m,
+        uint16 _n,
         uint64 _lotSizeSatoshis
     ) public payable returns (bool) {
         self.TBTCSystem = _TBTCSystem;
